bunch of refactoring weeee
