+++ packages/common/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "playhtml-common",
+  "private": true,
+  "license": "MIT",
+  "scripts": {},
+  "devDependencies": {
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9"
+  }
+}
