+++ packages/common/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src/index.tsxx
+  "include": ["**.ts"]
 }
