+++ packages/playhtml/src/main.ts
@@ -3,10 +3,10 @@
 import YPartyKitProvider from "y-partykit/provider";
 import { IndexeddbPersistence } from "y-indexeddb";
 import "./style.scss";
-import { ElementData, ElementInitializer, TagType } from "./types";
+import { ElementData, ElementInitializer, TagType } from "../../common/types";
 import * as Y from "yjs";
 import { TagTypeToElement, ElementHandler } from "./elements";
-import { getElementFromId, getIdForElement } from "./utils";
+import { getElementFromId, getIdForElement } from "../../common/utils";
 
 const DefaultPartykitHost = "playhtml.spencerc99.partykit.dev";
