+++ packages/playhtml/src/vite-env.d.ts
