+++ packages/react/example.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { Playable } from "./src/react";
+
+export function Candle() {
+  return <Playable />;
+}
