+++ packages/react/src/index.tsx
@@ -1,14 +1,7 @@
 import React, { PropsWithChildren, useEffect, useRef } from "react";
-import { useState, useCallback } from "react";
-import { PartykitHost, PartykitRoom, globalData } from "../src/main";
-import * as Y from "yjs";
-import { TagType } from "./types";
-import YPartyKitProvider from "y-partykit/provider";
-import { IndexeddbPersistence } from "y-indexeddb";
-import { useSyncedStore } from "@syncedstore/react";
-import { syncedStore, getYjsDoc } from "@syncedstore/core";
 import { playhtml } from "playhtml";
-import { ElementInitializer } from "playhtml/types";
+import { ElementInitializer, TagType } from "../../common/types";
+import { TagTypeToElement } from "../../playhtml/src/elements";
 
 // Get the Yjs document and sync automatically using y-webrtc
 
@@ -31,6 +24,8 @@ export function Playable(props: PropsWithChildren<ElementInitializer>) {
 }
 
 export function Movable(props: PropsWithChildren<{}>) {
-  return <Playable {...CanMoveProps}>{props.children}</Playable>;
+  return (
+    <Playable {...TagTypeToElement[TagType.CanMove]}>{props.children}</Playable>
+  );
 }
 //TODO: export an equivalent element for each of the things in elements.ts?
